chore(cicd): add build and start script
