+++ configs/dev.yml
@@ -1,0 +1,21 @@
+fs:
+  root: "tmp"
+  opensLimit: 128
+  openTTL: 60 # 1 min
+secrets:
+  tokenSecret: ""
+server:
+  debug: true
+  host: "0.0.0.0"
+  port: 8686
+  readTimeout: 2000
+  writerTimeout: 86400000 # 1 day
+  maxHeaderBytes: 512
+  publicPath: "public"
+users:
+  enableAuth: true
+  defaultAdmin: ""
+  defaultAdminPwd: ""
+  cookieTTL: 604800 # 1 week
+  cookieSecure: false
+  cookieHttpOnly: true
